# lib/api-zod/src/generated/types/createPlatformCustomerInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePlatformCustomerInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /** @pattern ^[a-z0-9][a-z0-9-]{2,62}$ */
  slug: string;
  /** @nullable */
  ownerEmail?: string | null;
}