# lib/api-zod/src/generated/types/createTenantInvitationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTenantInvitationInputRole } from './createTenantInvitationInputRole';

export interface CreateTenantInvitationInput {
  email: string;
  role: CreateTenantInvitationInputRole;
}