# lib/api-zod/src/generated/types/createdPlatformCustomer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformCustomer } from './platformCustomer';
import type { TenantInvitation } from './tenantInvitation';

export interface CreatedPlatformCustomer {
  customer: PlatformCustomer;
  invitation: TenantInvitation | null;
  /** @nullable */
  invitationToken: string | null;
}