# lib/api-zod/src/generated/types/createdTenantInvitation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TenantInvitation } from './tenantInvitation';

export interface CreatedTenantInvitation {
  invitation: TenantInvitation;
  token: string;
}