# lib/api-zod/src/generated/types/platformCustomerEnvironment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformCustomerEnvironmentKind } from './platformCustomerEnvironmentKind';
import type { PlatformCustomerEnvironmentStatus } from './platformCustomerEnvironmentStatus';

export interface PlatformCustomerEnvironment {
  id: number;
  name: string;
  kind: PlatformCustomerEnvironmentKind;
  status: PlatformCustomerEnvironmentStatus;
}